app/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0
django-crispy-forms==2.1
django-filter==23.5
sqlparse==0.4.4
